+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as InputBox } from './InputBox.js';
 export { default as Modal } from './Modal.js';
 export { default as Restaurant } from './Restaurant.js';
 export { default as RestaurantList } from './RestaurantList.js';
+export { default as RestaurantAddModal } from './RestaurantAddModal.js';
